pipeline_iptu_recife_projeto/iptu_recife/requirements.txt
# Core Data Analysis
pandas
pyarrow
fastparquet

# Cloud Storage and APIs
boto3
requests

# Environment Variables
python-dotenv